EPICS_CAS_IGNORE_NET_LIST lets servers ignore requests from whole subnets.
An EPICS channel access server will ignore requests from IP addresses that are
from one of the IP subnets specified in EPICS_CAS_IGNORE_NET_LIST.

The subnet specifications in EPICS_CAS_IGNORE_NET_LIST must be separated with
spaces.

Three forms of subnet specifications are supported:

- CIDR notation, example: 192.168.12.0/24
- Address and mask notation, example: 192.168.12.0:255.255.255.0
- Single IP address, example: 192.168.12.10

Notes for software that uses EPICS base:

This patch adds a private element to casDGIntfIO defined in casDGIntfIO.h.
It shouldn't be necessary to change the source code of applications
that are compiled against EPICS base, but they need to be recompiled.

In order to be able to detect if this patch is part of the currently used EPICS
base, a new C macro is defined in file epicsVersion.h:

  #define EPICS_HAS_CAS_IGNORE_NET_LIST 1